requirements.txt
# ML tool
numpy==1.19.1
scipy==1.5.2
pandas==1.1.3
openpyxl==3.0.7
scikit_learn==0.24.2
joblib==0.17.0
tqdm==4.51.0
omegaconf==2.1.0
scikit-multilearn==0.2.0
lightgbm==3.2.1

# data loading tool
mat73==0.52
pyampd==0.0.1

# logging tool
mlflow==1.18.0

# DL tool
torchaudio==0.8.0
pytorch-lightning==1.3.5
torchmetrics==0.6.0
protobuf==3.20.0
einops==0.6.0

# data visualization
coloredlogs==15.0
# hydra==1.1.0
hydra-optuna-sweeper==1.1.0
hydra-colorlog==1.1.0
